+++ src/java/org/apache/poi/hssf/record/formula/functions/Counta.java
@@ -34,7 +34,7 @@ import org.apache.poi.hssf.record.formula.functions.CountUtils.I_MatchPredicate;
  */
 public final class Counta implements Function {
 
-	public ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		int nArgs = args.length;
 		if (nArgs < 1) {
 			// too few arguments
