+++ src/java/org/apache/poi/hssf/record/formula/functions/Isref.java
@@ -27,7 +27,7 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
  * @author Amol S. Deshmukh &lt; amolweb at ya hoo dot com &gt;
  */
 public final class Isref implements Function {
-	public ValueEval evaluate(ValueEval[] operands, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] operands, int srcCellRow, int srcCellCol) {
 		if (operands.length != 1) {
 			return ErrorEval.VALUE_INVALID;
 		}
