+++ src/java/org/apache/poi/hssf/record/formula/functions/Row.java
@@ -25,7 +25,7 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
 
 public final class Row implements Function {
 
-    public ValueEval evaluate(ValueEval[] evals, int srcCellRow, short srcCellCol) {
+    public ValueEval evaluate(ValueEval[] evals, int srcCellRow, int srcCellCol) {
         ValueEval retval = null;
         int rnum = -1;
