+++ src/java/org/apache/poi/hssf/record/formula/functions/Vlookup.java
@@ -41,7 +41,7 @@ import org.apache.poi.hssf.record.formula.functions.LookupUtils.ValueVector;
  */
 public final class Vlookup implements Function {
 
-	public ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		ValueEval arg3 = null;
 		switch(args.length) {
 			case 4:
